Get rid of unnecessary allocation in succinctDuration
To create a duration with the most succint time unit typical usage
is "new Duration(value, unit).convertToMostSuccinctTimeUnit()" or
the succinctDuration method. However, both methods require an allocation
of a Duration instance and they can add up when the calling method is
hot. In Presto we observed that these account for ~3.5% of all
allocations on the coordinator. This change adds a private static
toSuccinctDuration method, which is used from succinctDuration, to get
rid of the unnecessary allocation.